Moved the client's chat history out of ChatStreamingWindow to a separate client helper component `ClientChatHistory` this holds the chat history and exposes a couple functions to add/clear history.
